+++ src/routes/(v2)/v2/allow-files/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { superForm } from '$lib/client/index.js';
 	import SuperDebug from '$lib/client/SuperDebug.svelte';
-	import { schema } from './schema.js';
+	//import { schema } from './schema.js';
 
 	export let data;